requirements.txt
astroid==2.5.1
autopep8==1.5.6
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dialogflow==1.1.0
dnspython==2.1.0
docx2txt==0.8
Flask==1.1.2
Flask-Cors==3.0.10
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-core==1.6.0
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
grpcio==1.36.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.5.2
libcst==0.3.18
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.3
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pylint==2.7.2
pymongo==3.11.3
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
